% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tengexa.r
\name{TEnGExA}
\alias{TEnGExA}
\title{Tissue-Enrichment Analysis of any Number of Genes and Tissues}
\usage{
TEnGExA(x, fpkm_flag = 0, threshold = 5, tissue_num = 5, min_fpkm = 1)
}
\arguments{
\item{x}{A numeric matrix or data frame of raw RNA-Seq read counts or fpkm values (file path).}

\item{fpkm_flag}{Numeric. 0 for count matrix or 1 for FPKM value matrix. Default-value 0}

\item{threshold}{Numeric. fpkm value threshold to be considered for calling any gene/transcript as expressed. Default-value 5}

\item{tissue_num}{Numeric. Minimum number of tissues to be considered for assigning group enrichment class. Default-value 5}

\item{min_fpkm}{Numeric. Minimum FPKM value threshold to be considered for expression analysis.  Default-value 1
# Default-output-file output1.csv}
}
\description{
Tissue-Enrichment Analysis of any Number of Genes and Tissues
}
